cgi work, chunked work, contentlen work
