+++ http/http.cpp
@@ -22,537 +22,6 @@
 #include "httpExceptions.h"
 #include <algorithm>
 
-const int64_t sessionTimeout = 2500000; // 2500 sec
-
-typedef void (HTTP::*handlerFunc)(int, Session*);
-
-struct Session
-{
-    void bind(handlerFunc f) {
-        func = f;
-    };
-
-    std::string   readBuf;
-    std::string   writeBuf;
-
-    handlerFunc func;
-
-    int32_t     fd;
-    int32_t     fdCGI;
-    size_t      size;
-    size_t      sizeChunked;
-    size_t      chunk;
-    Request     request;
-};
-
-HTTP::HTTP()
-    : sessionMap_()
-{
-
-}
-
-HTTP::~HTTP()
-{
-
-}
-
-void HTTP::start() {
-    EventPool::start();
-}
-
-void HTTP::listen(const std::string& host)
-{
-    try {
-        TcpSocket conn(host);
-        conn.listen();
-        conn.makeNonBlock();
-
-        Session listenSession;
-        newListenerEvent(conn);
-        newSessionByID(conn.getSock(), listenSession);
-    }  catch (std::exception& e) {
-        LOG_ERROR("HTTP: %s\n", e.what());
-    }
-}
-
-Session* HTTP::getSessionByID(int id)
-{
-    if (sessionMap_.find(id) != sessionMap_.end()) {
-        return &sessionMap_[id];
-    }
-    return nullptr;
-}
-
-void HTTP::closeSessionByID(int id) {
-    disableTimerEvent(id, 0);
-    ::close(id);
-    sessionMap_.erase(id);
-}
-
-void HTTP::newSessionByID(int id, Session& session)
-{
-    sessionMap_[id] = session;
-}
-
-////////////////////////////////////////////////////////////////////////////////
-////////////////////////////// ASYNC LOGIC /////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////
-
-void HTTP::asyncAccept(TcpSocket& socket)
-{
-    LOG_DEBUG("Event accepter call\n");
-
-    try
-    {
-        Session *session = getSessionByID(socket.getSock());
-        if (session == nullptr) {
-            LOG_ERROR("Dont find session to socket: %d. Close.\n", socket.getSock());
-            close(socket.getSock());
-            return;
-        }
-
-        TcpSocket conn = socket.accept();
-
-        LOG_DEBUG("New connect fd: %d\n", conn.getSock());
-
-        Session s;
-        s.bind(&HTTP::defaultReadFunc);
-        s.request.setHost(conn.getHost());
-        s.request.setID(conn.getSock());
-
-        newSessionByID(conn.getSock(), s);
-        enableReadEvent(conn.getSock());
-        enableTimerEvent(conn.getSock(), sessionTimeout); // TODO add config
-    }
-    catch (std::exception& e)
-    {
-        LOG_ERROR("HTTP::asyncAccept error: %s\n", e.what());
-    }
-};
-
-void HTTP::asyncWrite(int socket)
-{
-    LOG_DEBUG("Event writer call\n");
-
-    Session *session = getSessionByID(socket);
-    if (session == nullptr) {
-        LOG_ERROR("Dont find session to socket: %d. Close connection.\n", socket);
-        closeSessionByID(socket);
-        return;
-    }
-    (this->*session->func)(socket, session);
-}
-
-void HTTP::asyncRead(int socket)
-{
-    LOG_DEBUG("Event reader call\n");
-
-    Session *session = getSessionByID(socket);
-    if (session == nullptr) {
-        LOG_ERROR("Dont find session to socket: %d. Close connection.\n", socket);
-        closeSessionByID(socket);
-        return;
-    }
-    (this->*session->func)(socket, session);
-}
-
-void HTTP::asyncEvent(int socket, uint16_t flags)
-{
-    LOG_DEBUG("Event handler call\n");
-    if (flags & EventPool::M_EOF || flags & EventPool::M_ERROR) {
-        LOG_ERROR("Event error or eof, socket: %d\n", socket);
-        closeSessionByID(socket);
-    }
-    if (flags & EventPool::M_TIMER) {
-        LOG_ERROR("Event timeout, socket: %d\n", socket);
-        closeSessionByID(socket);
-    }
-}
-
-static ssize_t readToBuf(int fd, std::string& rBuf)
-{
-    const size_t bufSize = (1 << 16);
-    char buf[bufSize];
-
-    ssize_t readBytes = ::read(fd, buf, bufSize);
-    if (readBytes < 0) {
-        return readBytes;
-    }
-
-    rBuf.append(buf, buf + readBytes);
-
-    return readBytes;
-}
-
-static ssize_t writeFromBuf(int fd, std::string& wBuf, size_t nBytes)
-{
-    ssize_t writeBytes = ::write(fd, wBuf.c_str(), nBytes);
-    if (writeBytes < 0) {
-        return writeBytes;
-    }
-
-    wBuf.erase(0, writeBytes);
-
-    return writeBytes;
-}
-
-
-void HTTP::defaultReadFunc(int socket, Session *session)
-{
-    LOG_DEBUG("defaultReadFunc call\n");
-
-    std::string& rbuf = session->readBuf;
-
-    if (readToBuf(socket, rbuf) < 0) {
-        closeSessionByID(socket);
-        LOG_ERROR("Dont read to socket: %d. Close connection.\n", socket);
-        return;
-    }
-
-    size_t foundNN = rbuf.find("\n\n");
-    size_t foundRN = rbuf.find("\r\n\r\n");
-
-    size_t found = std::min<size_t>(foundNN, foundRN);
-    if (found != std::string::npos) {
-        size_t pos = rbuf.find_first_not_of("\r\n", found);
-        session->request.reset();
-        session->request.parse(std::string(rbuf, 0, pos));
-        rbuf.erase(0, pos);
-
-        session->bind(&HTTP::defaultWriteFunc);
-		LOG_DEBUG("REDIRECTION IN defaultReadFunction line208");
-        // включаем write
-        enableWriteEvent(socket);
-        disableReadEvent(socket);
-    }
-}
-
-void HTTP::defaultWriteFunc(int socket, Session *session)
-{
-    LOG_DEBUG("defaultWriteFunc call\n");
-    session->bind(&HTTP::doneFunc);
-
-    Request& request = session->request;
-
-    Response response;
-    handler(request, response);
-
-    std::string& wbuf = session->writeBuf;
-	LOG_DEBUG("REDIRECTION HERE\n");
-	std::cout << response.getContent();
-
-    wbuf.append(response.getContent());
-}
-
-void HTTP::doneFunc(int socket, Session *session)
-{
-    LOG_DEBUG("doneFunc call\n");
-    std::string& wbuf = session->writeBuf;
-
-    if (wbuf.empty()) {
-        disableWriteEvent(socket);
-        enableReadEvent(socket);
-        session->bind(&HTTP::defaultReadFunc);
-    }
-    else {
-        if (writeFromBuf(socket, wbuf, wbuf.size()) < 0) {
-            closeSessionByID(socket);
-            LOG_ERROR("Dont write to socket: %d. Close connection.\n", socket);
-            return;
-        }
-    }
-}
-
-void HTTP::sendFileFunc(int socket, Session* session)
-{
-    LOG_DEBUG("sendFileFunc call\n");
-
-    std::string& wbuf = session->writeBuf;
-
-    ssize_t readBytes = readToBuf(session->fd, wbuf);
-    if (readBytes < 0) {
-        close(session->fd);
-        closeSessionByID(socket);
-        LOG_ERROR("Dont read to file: %d. Close connection.\n", socket);
-        return;
-    }
-
-    size_t bytes = std::min<size_t>(wbuf.size(), session->size);
-    ssize_t writeBytes = writeFromBuf(socket, wbuf, bytes);
-    if (writeBytes < 0) {
-        close(session->fd);
-        closeSessionByID(socket);
-        LOG_ERROR("Dont write to socket: %d. Close connection.\n", socket);
-        return;
-    }
-
-    session->size -= writeBytes;
-
-    if (session->size == 0 || readBytes == 0) {
-        close(session->fd);
-        session->bind(&HTTP::doneFunc);
-    }
-}
-
-void HTTP::recvFileFunc(int socket, Session* session)
-{
-    LOG_DEBUG("recvFileFunc call\n");
-
-    std::string& rbuf = session->readBuf;
-
-    if (readToBuf(socket, rbuf) < 0) {
-        closeSessionByID(socket);
-        LOG_ERROR("Dont read to file: %d. Close connection.\n", socket);
-        return;
-    }
-
-    size_t bytes = std::min<size_t>(rbuf.size(), session->size);
-
-    ssize_t writeBytes = writeFromBuf(session->fd, rbuf, bytes);
-
-    if (writeBytes < 0) {
-        close(session->fd);
-        closeSessionByID(socket);
-        LOG_ERROR("Dont write to socket: %d. Close connection.\n", socket);
-        return;
-    }
-
-    session->size -= writeBytes;
-    if (session->size == 0) {
-        session->bind(&HTTP::doneFunc);
-        disableReadEvent(socket);
-        enableWriteEvent(socket);
-    }
-}
-
-void HTTP::recvCGICallback(int socket, Session* session)
-{
-    LOG_DEBUG("recvFileFunc call\n");
-
-    std::string& rbuf = session->readBuf;
-
-    if (readToBuf(socket, rbuf) < 0) {
-        closeSessionByID(socket);
-        LOG_ERROR("Dont read to file: %d. Close connection.\n", socket);
-        return;
-    }
-
-    size_t bytes = std::min<size_t>(rbuf.size(), session->size);
-
-    ssize_t writeBytes = writeFromBuf(session->fd, rbuf, bytes);
-
-    if (writeBytes < 0) {
-        close(session->fd);
-        close(session->fdCGI);
-        closeSessionByID(socket);
-        LOG_ERROR("Dont write to socket: %d. Close connection.\n", socket);
-        return;
-    }
-
-    session->size -= writeBytes;
-    if (session->size == 0) {
-        session->bind(&HTTP::cgiCaller);
-        disableReadEvent(socket);
-        enableWriteEvent(socket);
-        return;
-    }
-}
-
-static int32_t writeChunk(std::string& rbuf, Session* session) {
-    while (!rbuf.empty()) {
-
-        if (session->chunk == 0) {
-            rbuf.erase(0, rbuf.find_first_not_of("\r\n"));
-            size_t found = rbuf.find("\n");
-            if (found != std::string::npos) {
-                size_t pos = rbuf.find_first_not_of("\r\n", found);
-                session->chunk = utils::to_number<size_t>(std::string(rbuf, 0, pos));
-                rbuf.erase(0, pos);
-            }
-            else {
-                return 1;
-            }
-        }
-
-
-        if (session->chunk == 0) {
-            return 0;
-        }
-
-
-        while (!rbuf.empty() && session->chunk > 0) {
-            ssize_t writeBytes = writeFromBuf(session->fd, rbuf, session->chunk);
-            if (writeBytes < 0) {
-                return -1;
-            }
-
-            session->chunk -= writeBytes;
-            session->sizeChunked += writeBytes;
-            if (session->sizeChunked > session->size) {
-                return -1;
-            }
-        }
-    }
-    return 1;
-}
-
-void HTTP::recvChunkedToFileCallback(int socket, Session *session)
-{
-    std::string& rbuf = session->readBuf;
-
-    if (readToBuf(socket, rbuf) < 0) {
-        closeSessionByID(socket);
-        LOG_ERROR("Dont read to file: %d. Close connection.\n", socket);
-        return;
-    }
-
-    int32_t ret = writeChunk(rbuf, session);
-
-    if (ret < 0) {
-        close(session->fd);
-        closeSessionByID(socket);
-        LOG_ERROR("Dont write to fd: %d. Close connection.\n", socket);
-    }
-    if (ret == 0) {
-        close(session->fd);
-        session->bind(&HTTP::doneFunc);
-        enableWriteEvent(socket);
-        disableReadEvent(socket);
-    }
-}
-
-void HTTP::recvChunkedToCGICallback(int socket, Session *session)
-{
-    std::string& rbuf = session->readBuf;
-
-    if (readToBuf(socket, rbuf) < 0) {
-        closeSessionByID(socket);
-        LOG_ERROR("Dont read to file: %d. Close connection.\n", socket);
-        return;
-    }
-
-    int32_t ret = writeChunk(rbuf, session);
-
-    if (ret < 0) {
-        close(session->fd);
-        close(session->fdCGI);
-        closeSessionByID(socket);
-        LOG_ERROR("Dont write to fd: %d. Close connection.\n", socket);
-    }
-    if (ret == 0) {
-        close(session->fd);
-        session->bind(&HTTP::cgiCaller);
-        enableWriteEvent(socket);
-        disableReadEvent(socket);
-    }
-}
-
-
-
-
-/*
-void HTTP::recvChunkedWriteToFile(int socket, Session* session)
-{
-
-    std::string& rbuf = session->readBuf;
-
-    if (rbuf.empty() && session->chunkSize == 0) {
-        session->bind(&HTTP::recvChunkedReadFromSocket);
-        enableReadEvent()
-    }
-
-    if (readToBuf(socket, rbuf) < 0) {
-        closeSessionByID(socket);
-        LOG_ERROR("Dont read to file: %d. Close connection.\n", socket);
-        return;
-    }
-
-    if (session->size == 0) {
-        rbuf.erase(0, rbuf.find_first_not_of("\r\n"));
-        disableReadEvent(socket);
-        enableWriteEvent(socket);
-        session->bind(&HTTP::doneFunc);
-    }
-
-    if (rbuf.size() > size) {
-
-
-
-        while (size > 0) {
-            size_t writeBytes = ::write(fd, rbuf.c_str(), size);
-
-            if (writeBytes < 0) {
-                close(fd);
-                closeSessionByID(socket);
-                LOG_ERROR("Dont write to fd: %d. Close connection.\n", socket);
-                return;
-            }
-
-            rbuf.erase(0, writeBytes);
-            size -= writeBytes;
-        }
-
-        rbuf.erase(0, rbuf.find_first_not_of("\r\n"));
-        session->bind(&HTTP::recvChunkedReadFromSocket);
-    }
-
-}
-
-void HTTP::recvChunkedWriteToSocket(int socket, Session* session)
-{
-
-}
-*/
-void HTTP::sendChunkedCallback(int socket, Session *session)
-{
-
-    const int32_t bufSize = (1 << 16);
-    char buf[bufSize] = {};
-
-    int32_t fd = session->fd;
-    int32_t readBytes = ::read(fd, buf, bufSize);
-
-    if (readBytes < 0) {
-        close(fd);
-        closeSessionByID(socket);
-        LOG_ERROR("Dont read to file: %d. Close connection.\n", socket);
-        return;
-    }
-
-    std::string& wbuf = session->writeBuf;
-    wbuf.append(utils::to_string<int32_t>(readBytes));
-    wbuf.append("\r\n");
-    wbuf.append(buf, buf + readBytes);
-    wbuf.append("\r\n");
-
-    size_t writeBytes = writeFromBuf(socket, wbuf, wbuf.size());
-
-    if (writeBytes < 0) {
-        close(fd);
-        closeSessionByID(socket);
-        LOG_ERROR("Dont write to socket: %d. Close connection.\n", socket);
-        return;
-    }
-
-    if (readBytes == 0) {
-        session->bind(&HTTP::doneFunc);
-        close(fd);
-    }
-
-}
-
-void HTTP::cgiCaller(int socket, Session* session)
-{
-    LOG_DEBUG("CgiCaller call\n");
-    Request& request = session->request;
-
-    std::string s;
-    readToBuf(session->fdCGI, s);
-
-    session->writeBuf.append(s);
-    session->bind(&HTTP::doneFunc);
-}
-
 ///////////////////////////////////////////////////////////////////////////
 ///////////////////////// HTTP LOGIC //////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////
@@ -560,15 +29,14 @@ void HTTP::cgiCaller(int socket, Session* session)
 bool HTTP::cgi(const Request &request, Response& response, Route* route) {
     const std::string& path = request.getPath();
     bool isCGI = route != nullptr && utils::getExtension(path) == route->getCgi();
-    if (isCGI) {
-		std::string body;
+    if (isCGI)
+    {
         if (!utils::isFile(route->getFullPath(path)))
             throw httpEx<NotFound>("CGI script not found");
-        body = Cgi(request, *route).runCGI();
-        //        response.setHeaderField("Content-Type", request.getHeaders().at("Content-Type"));
-        response.setHeaderField("Content-Length", body.size());
-        response.setStatusCode(200);
-        response.setBody(body);
+
+        int fd = Cgi(request, *route).runCGI(request.fd);
+
+        sendFile(request, response, fd, true);
     }
     return isCGI;
 }
@@ -624,22 +92,11 @@ void HTTP::handler(Request& request, Response& response) {
 		}
 
 
-        //recvCGI(request, response);
-        //recvCGIChunked(request, response);
-        //sendFileChunked(request, response, "./test.sh");
-        //recvFileChunked(request, response, "./test_chunked");
+
+        //sendFile(request, response, "./test.sh", true);
+        //saveToFile(request.fd, "./my_file");
         //return;
 
-
-        //if (request.getMethod() == "GET") {
-        //    sendFile(request, response, "./index.html");
-        //    return;
-        //}
-        //if (request.getMethod() == "POST") {
-        //    recvFile(request, response, "./my_file");
-        //    return;
-        //}
-        //return;
 
         if (route == nullptr) {
             throw httpEx<NotFound>("Not Found");
@@ -827,7 +284,8 @@ void HTTP::methodPOST(const Request& request, Response& response, Route* route){
     std::cout << request << std::endl;
     const std::string& path = route->getFullPath(request.getPath());
     const std::string& body = request.getBody();
-    response.writeFile(path, body);
+
+
     response.setStatusCode(201);
     response.setHeaderField("Content-Location", "/filename.xxx");
 }
@@ -941,213 +399,6 @@ void HTTP::startServer()
     }
     serve.start();
 }
-
-
-void HTTP::sendFile(Request& request, Response& response, const std::string& path)
-{
-    int fd = ::open(path.c_str(), O_RDONLY);
-
-    if (fd == -1) {
-        throw httpEx<InternalServerError>("Cannot open file: " + path);
-    }
-
-    struct stat info;
-
-    if (::fstat(fd, &info) == -1) {
-        throw httpEx<InternalServerError>("Cannot stat file info: " + path);
-    }
-
-    response.setStatusCode(200);
-    response.setHeaderField("Content-Length", info.st_size);
-    response.setContent(response.getHeader());
-
-    Session* session = getSessionByID(request.getID());
-
-    session->fd = fd;
-    session->size = info.st_size;
-    session->bind(&HTTP::sendFileFunc);
-}
-
-void HTTP::recvFile(Request& request, Response& response, const std::string& path)
-{
-    int fd = ::open(path.c_str(), O_WRONLY|O_CREAT, 0664);
-
-    if (fd == -1) {
-        throw httpEx<InternalServerError>("Cannot open file: " + path);
-    }
-
-    response.setStatusCode(200);
-
-    Session* session = getSessionByID(request.getID());
-
-    if (request.hasHeader("Content-Length")) {
-        session->size = utils::to_number<size_t>(request.getHeaderValue("Content-Length"));
-    }
-    else {
-        LOG_INFO("HTTP request 'Content-Length' header not found!\n");
-        session->size = 0;
-    }
-
-    std::string& rbuf = session->readBuf;
-
-    int64_t bytes = std::min<int64_t>(rbuf.size(), session->size);
-
-
-    while (bytes > 0) {
-        ssize_t writeBytes = writeFromBuf(session->fd, rbuf, bytes);
-
-        if (writeBytes < 0) {
-            close(fd);
-            closeSessionByID(request.getID());
-            LOG_ERROR("Dont write to socket: %d. Close connection.\n", socket);
-            return;
-        }
-        bytes -= writeBytes;
-        session->size -= writeBytes;
-    }
-
-    if (session->size == 0) {
-        return;
-    }
-
-
-    session->bind(&HTTP::recvFileFunc);
-    enableReadEvent(request.getID());
-    disableWriteEvent(request.getID());
-}
-
-void HTTP::recvFileChunked(Request& request, Response& response, const std::string& path)
-{
-    int fd = ::open(path.c_str(), O_WRONLY|O_CREAT, 0664);
-
-    if (fd == -1) {
-        throw httpEx<InternalServerError>("Cannot open file: " + path);
-    }
-
-    Session* session = getSessionByID(request.getID());
-
-    session->fd = fd;
-    session->size = 0;
-    session->chunk = 0;
-    session->sizeChunked = 0;
-
-    int ret = writeChunk(session->readBuf, session);
-
-    if (ret < 0) {
-        close(fd);
-        closeSessionByID(request.getID());
-    }
-    if (ret == 0) {
-        close(fd);
-        return;
-    }
-
-    session->bind(&HTTP::recvChunkedToFileCallback);
-    enableReadEvent(request.getID());
-    disableWriteEvent(request.getID());
-}
-
-void HTTP::recvCGIChunked(Request& request, Response& response)
-{
-    int fds[2];
-
-    if (::pipe(fds) == -1) {
-        throw httpEx<InternalServerError>("HTTP::Cannot open pipe!");
-    }
-
-    Session* session = getSessionByID(request.getID());
-
-    session->fd = fds[1];
-    session->fdCGI = fds[0];
-    session->size = 1000;
-    session->chunk = 0;
-    session->sizeChunked = 0;
-
-    int ret = writeChunk(session->readBuf, session);
-
-    if (ret < 0) {
-        close(fds[0]);
-        close(fds[1]);
-        closeSessionByID(request.getID());
-    }
-    if (ret == 0) {
-        close(fds[1]);
-        session->bind(&HTTP::cgiCaller);
-        return;
-    }
-
-    session->bind(&HTTP::recvChunkedToCGICallback);
-    enableReadEvent(request.getID());
-    disableWriteEvent(request.getID());
-}
-
-void HTTP::recvCGI(Request& request, Response& response)
-{
-    int fds[2];
-
-    if (::pipe(fds) == -1) {
-        throw httpEx<InternalServerError>("HTTP::Cannot open pipe!");
-    }
-
-    Session* session = getSessionByID(request.getID());
-
-    session->fd = fds[1];
-    session->fdCGI = fds[0];
-
-    if (request.hasHeader("Content-Length")) {
-        session->size = utils::to_number<size_t>(request.getHeaderValue("Content-Length"));
-    }
-    else {
-        LOG_INFO("HTTP request 'Content-Length' header not found!\n");
-        session->size = 0;
-    }
-
-    std::string& rbuf = session->readBuf;
-    size_t bytes = std::min<size_t>(rbuf.size(), session->size);
-
-    ssize_t writeBytes = writeFromBuf(session->fd, rbuf, bytes);
-
-    if (writeBytes < 0) {
-        close(session->fd);
-        close(session->fdCGI);
-        closeSessionByID(request.getID());
-        LOG_ERROR("Dont write to socket: %d. Close connection.\n", socket);
-        return;
-    }
-
-    session->size -= writeBytes;
-    if (session->size == 0) {
-        session->bind(&HTTP::cgiCaller);
-        disableReadEvent(request.getID());
-        enableWriteEvent(request.getID());
-        return;
-    }
-
-
-    session->bind(&HTTP::recvCGICallback);
-    enableReadEvent(request.getID());
-    disableWriteEvent(request.getID());
-}
-
-void HTTP::sendFileChunked(Request& request, Response& response, const std::string& path)
-{
-    int fd = ::open(path.c_str(), O_RDONLY);
-
-    if (fd == -1) {
-        throw httpEx<InternalServerError>("Cannot open file: " + path);
-    }
-
-    Session* session = getSessionByID(request.getID());
-
-    session->fd = fd;
-    session->size = 0;
-    session->chunk = 0;
-    session->sizeChunked = 0;
-
-    session->bind(&HTTP::sendChunkedCallback);
-}
-
-
 
 bool HTTP::isValidMethod(const std::string &method)
 {
